zsh: Move path settings to `.zprofile`
MacOS uses `/etc/zprofile` instead of `/etc/zshenv` to set the default
path data. That gets called after `~/.zshenv` which ends up prepending
paths like `/bin` and `/usr/bin` to the front of the list. I have to
move my path configuration to `~/.zprofile` so that it runs after the
default paths are set allowing me to prepend my paths.