Update the installer to actually work correctly
The "sublime" and ".config" folders need special handling, and the
"_docs" folder doesn't need to be linked at all. This updates the
installer to correctly handle things. (Though there's room to improve
its robustness.)

Also added a symlink to the installer named "installer.sh" for some
automated tooling that looks for ".sh" instead of ".zsh"