starship: Fix the git status markers for Linux
In MacOS, the "-E" option seems to use PCRE, but for Linux its the same
as the basic "-e" option, and Linux requires using "-P" to get actual
regular expression extended from the basic expressions. Rather than
trying to figure out how to swap between "-E" and "-P", I'm modifying
the regular expression to not bother with ignoring the capture group
("?:"). This gets it working in Linux as well as MacOS.